Implement geometry class and add geometry set
The geometry class represents an object inside the fluid. It will be
created by a reference to a file parser. It copies all points in that
file to its data storage. It stops adding points when the termination
symbol is reached (see commit 41cf5b56902a3650bac987b6990f294eb6e3582f).

A geometry set is a container for a bunch of geometries. Given a
filepath to a .fsg file it adds all of the contained objects into its
own data storage by creating instances of the geometry class.