Add boundary recognition for geometries
While parsing a file and creating a geometry, the object no keeps track
of its minimum boundary frame. This will later be used to automatically
create an adequate mesh for the calculation of the fluid flow.

Additionally, a bug occurring during the instantiation of multiple
geometries was fixed. The end of the parsed file was not recognised in
time and thus, a geometry was created even though it contained no
points. The origin of the bug was located in the file parser for
geometries.