Outsource boundary calculations from geometry
The minimal boundary calculation for geometries was outsourced so that
the geometry set and later coming classes can do similar calculations
as this will be widely used algorithm.

With the same reasoning, a point data structure was created.

The Utility component was created. Minor algorithms like the boundary
calculation and minor data structures like a point in space can be
implemented there. The Geometry component depends on it. It is expected
that most of the other components will depend on it.