Add graphics and make geometry renderable
A new component, Graphics, was created to control all of the visual
output of the program. The component depends on SFML, the library
chosen for rendering and all components containing the content to
visualize. Classes with renderable data should inherit from the class
Renderable class in Graphics. To prevent that other components depend
on SFML, all graphics files should be contained in the Graphics
component.

The .fsg file format was extended to give more information about
represented geometries. The last line for any object having the point
number 0 should now also state whether the line is closed or open. This
is done by setting the x-coordinate to 1 or 0, respectively.

The Geometry and Utility component were extended with more
functionality to make gathering data for rendering easier and better
encapsulated.