Create mesh generation and centralize graphics
SFML was completely outsourced to the Graphics component. Its new class,
the RenderController now possesses the window and all renderable
entities, like the previously implemented RenderableGeometry. For the
new mesh generator a renderable class was also created.

To make frequent graphics operations more available, the RenderUtility
class was created. One object will be instantiated from the singleton
RenderController. It will be passed to all renderable objects, so they
can perform for example coordinate transformations.

A new component, Mesh, was created. It consists of a MeshCell and a mesh
class. After a geometry was created, the mesh surrounding the object of
interest will be created by this class. For now, it will create a
cartesian mesh suitable for the finite difference method.

Finding out, whether a cell is inside the geometry is an important task
here. Or in an abstract form: It has to be determined if a point is
inside a polygon. For this problem, the line structure was created and
the boundary structure was extended.